refactor: rename maps components (#1246)
