+++ sites/geohub/src/components/maps/MapStyleCard.svelte
